https://github.com/Tour-de-Force/btc-infrastructure/issues/1
Set USERPROFILE environment variable to a directory outside System32. When the Workgroup/System user was running Chef and had its home folder inside System32, node-gyp was getting sad. Node-gyp attempts to build native dependencies and was keeping some config data in USERPROFILE. It was the thing supposedly failing on Python not being installed. When its USERPROFILE is outside System32 it seems to be able to handle itself.

However, the dependency that needs node-gyp (leveldown depended on by the PouchDB package) is still failing to build for another reason, but it is an optional dependency. This failure was causing a nonzero return from the Powershell script, so we can force "exit 0" from that script until we fix the dependency. We will just have to check that the server is really running after deploying in the meantime. I will write a **new bug** for this build fix if we're okay with how it is for now.

Also set the npm prefix and cache paths with environment variables instead of depending on it finding the right config files with all this System32 nonsense going on. Environment variables are the highest precedence for npm too; so we can be more sure they will be honored.

I was able to deploy 2 new web servers and 2 new database servers with this patch.